--- begin-react/src/App.js
@@ -1,17 +1,0 @@
-import logo from './logo.svg';
-import './App.css';
-import Calendar from './calendar.jsx';
-import TotalGi from './TotalGi.jsx';
-import Pick from './foodpick.jsx';
-import Check from './checklist.jsx';
-
-const App = () => {
-  return (
-    <div>
-      <h1>Main App Component</h1>
-      <TotalGi></TotalGi>
-    </div>
-  );
-};
-
-export default App;
