+++ src/index.js
@@ -14,6 +14,14 @@ import Developer from './Developer';
 import DoctorSignUp from './DoctorSignUp';
 import ChatRoom from './ChatRoom';
 import Question from './Question';
+import Calendar from './calendar';
+import TotalGi from './totalGi';
+import Pick0 from './foodpick0';
+import Pick1 from './foodpick1';
+import Pick2 from './foodpick2';
+import Pick3 from './foodpick3';
+import CheckList from './checklist/checklist';
+
 
 const Router = () => {
   return(
@@ -26,6 +34,14 @@ const Router = () => {
         <Route path='/developerIntroduction' element={<Developer></Developer>}></Route>
         <Route path='/doctorsignup' element={<DoctorSignUp/>}></Route>
         <Route path='/question' element={<Question></Question>}></Route>
+        <Route path='/home' element={<Calendar></Calendar>}></Route>
+        <Route path='/gi' element={<TotalGi></TotalGi>}></Route>
+        <Route path='/Pick0' element={<Pick0></Pick0>}></Route>
+        <Route path='/Pick1' element={<Pick1></Pick1>}></Route>
+        <Route path='/Pick2' element={<Pick2></Pick2>}></Route>
+        <Route path='/Pick3' element={<Pick3></Pick3>}></Route>
+        <Route path='checklist' element={<CheckList></CheckList>}></Route>
+        
 
       </Routes>
